componentize reader
